Move Kotlin task creation tests from GradleKotlinDslIntegrationTest to TaskDefinitionIntegrationSpec
