+++ subprojects/integ-test/src/integTest/groovy/org/gradle/integtests/GradleKotlinDslIntegrationTest.groovy
@@ -20,7 +20,6 @@ import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.test.fixtures.file.LeaksFileHandles
 import org.gradle.test.fixtures.server.http.HttpServer
 import org.gradle.util.Requires
-import spock.lang.Ignore
 
 import static org.gradle.util.TestPrecondition.KOTLIN_SCRIPT
 
@@ -54,137 +53,6 @@ class GradleKotlinDslIntegrationTest extends AbstractIntegrationSpec {
 
         then:
         result.output.contains('it works!')
-    }
-
-    def 'can run a custom task with constructor arguments via API'() {
-        given:
-        buildFile << """
-            import org.gradle.api.*
-            import org.gradle.api.tasks.*
-            import javax.inject.Inject
-
-            open class CustomTask @Inject constructor(private val message: String, private val number: Int) : DefaultTask() {
-                @TaskAction fun run() = println("\$message \$number")
-            }
-
-            tasks.create("myTask", CustomTask::class.java, "hello", 42)
-        """
-
-        when:
-        run 'myTask'
-
-        then:
-        result.output.contains('hello 42')
-    }
-
-    @Ignore
-    def 'can run custom task with constructor arguments via extension'() {
-        given:
-        buildFile << """
-            import org.gradle.api.*
-            import org.gradle.api.tasks.*
-            import javax.inject.Inject
-
-            open class CustomTask @Inject constructor(private val message: String, private val number: Int) : DefaultTask() {
-                @TaskAction fun run() = println("\$message \$number")
-            }
-
-            task<CustomTask>("myTask", "hello", 42)
-        """
-
-        when:
-        run 'myTask'
-
-        then:
-        result.output.contains('hello 42')
-    }
-
-    def "fails to build custom task if constructor arguments missing"() {
-        given:
-        buildFile << """
-            import org.gradle.api.*
-            import org.gradle.api.tasks.*
-            import javax.inject.Inject
-
-            open class CustomTask @Inject constructor(private val message: String, private val number: Int) : DefaultTask() {
-                @TaskAction fun run() = println("\$message \$number")
-            }
-
-            tasks.create("myTask", CustomTask::class.java, "hello")
-        """
-
-        when:
-        fails 'myTask'
-
-        then:
-        result.output.contains("org.gradle.internal.service.UnknownServiceException: No service of type int available")
-    }
-
-    def "can construct a task with @Inject services"() {
-        given:
-        buildFile << """
-            import org.gradle.api.*
-            import org.gradle.api.tasks.*
-            import org.gradle.workers.WorkerExecutor
-            import javax.inject.Inject
-
-            open class CustomTask @Inject constructor(private val executor: WorkerExecutor) : DefaultTask() {
-                @TaskAction fun run() = println(if (executor != null) "got it" else "NOT IT")
-            }
-
-            task<CustomTask>("myTask")
-        """
-
-        when:
-        run 'myTask'
-
-        then:
-        result.output.contains("got it")
-    }
-
-    def "can construct a task with @Inject services and constructor args"() {
-        given:
-        buildFile << """
-            import org.gradle.api.*
-            import org.gradle.api.tasks.*
-            import org.gradle.workers.WorkerExecutor
-            import javax.inject.Inject
-
-            open class CustomTask @Inject constructor(private val number: Int, private val executor: WorkerExecutor) : DefaultTask() {
-                @TaskAction fun run() = println(if (executor != null) "got it \$number" else "\$number NOT IT")
-            }
-
-            tasks.create("myTask", CustomTask::class.java, 15)
-        """
-
-        when:
-        run 'myTask'
-
-        then:
-        result.output.contains("got it 15")
-    }
-
-    @Ignore
-    def "can construct a task with @Inject services and constructor args via extension"() {
-        given:
-        buildFile << """
-            import org.gradle.api.*
-            import org.gradle.api.tasks.*
-            import org.gradle.workers.WorkerExecutor
-            import javax.inject.Inject
-
-            open class CustomTask @Inject constructor(private val number: Int, private val executor: WorkerExecutor) : DefaultTask() {
-                @TaskAction fun run() = println(if (executor != null) "got it \$number" else "\$number NOT IT")
-            }
-
-            tasks<CustomTask>("myTask", 15)
-        """
-
-        when:
-        run 'myTask'
-
-        then:
-        result.output.contains("got it 15")
     }
 
     @LeaksFileHandles
